move test case to tests/specs as NPM_CONFIG_REGISTRY is not set in js unit test
