+++ tests/specs/run/resolve_dns_edns0/test.out
@@ -1,0 +1,6 @@
+[WILDCARD]
+running 1 test from ./test.ts
+EDNS0 enabled ... ok ([WILDCARD]s)
+
+ok | 1 passed | 0 failed ([WILDCARD]s)
+
